Fix avg_d

+++ realworld_benchmark/main_PCBA.py
@@ -386,7 +386,7 @@ def main():
     if args.not_pre is not None:
         net_params['not_pre'] = args.not_pre
 
-    D = torch.cat([dataset.train[i][0].number_of_nodes() for i in range(len(dataset.train))])
+    D = torch.cat([torch.FloatTensor(dataset.train[i][0].number_of_nodes()) for i in range(len(dataset.train))])
     net_params['avg_d'] = dict(lin=torch.mean(D),
                                exp=torch.mean(torch.exp(torch.div(1, D)) - 1),
                                log=torch.mean(torch.log(D + 1)))
