+++ realworld_benchmark/main_superpixels_hydra.py
@@ -448,7 +448,7 @@ def main():
     num_classes = len(np.unique(np.array(dataset.train[:][1])))
     net_params['n_classes'] = num_classes
 
-    D = torch.cat([dataset.train[i][0].number_of_nodes() for i in range(len(dataset.train))])
+    D = torch.cat([torch.FloatTensor(dataset.train[i][0].number_of_nodes()) for i in range(len(dataset.train))])
     net_params['avg_d'] = dict(lin=torch.mean(D),
                                exp=torch.mean(torch.exp(torch.div(1, D)) - 1),
                                log=torch.mean(torch.log(D + 1)))
